+++ src/Oras.Cli/Commands/CopyCommand.cs
@@ -1,4 +1,5 @@
 using System.CommandLine;
+using Microsoft.Extensions.DependencyInjection;
 using Oras.Options;
 using Oras.Services;
 using Oras.Output;
@@ -73,8 +74,7 @@ internal static class CopyCommand
         {
             return await ErrorHandler.HandleAsync(async () =>
             {
-                var registryService = serviceProvider.GetService(typeof(IRegistryService)) as IRegistryService
-                    ?? throw new InvalidOperationException("Registry service not available");
+                var registryService = serviceProvider.GetRequiredService<IRegistryService>();
 
                 var source = parseResult.GetValue(sourceArg)!;
                 var destination = parseResult.GetValue(destArg)!;