+++ src/Oras.Cli/Commands/PullCommand.cs
@@ -1,4 +1,5 @@
 using System.CommandLine;
+using Microsoft.Extensions.DependencyInjection;
 using Oras.Options;
 using Oras.Services;
 using Spectre.Console;
@@ -52,8 +53,7 @@ internal static class PullCommand
         {
             return await ErrorHandler.HandleAsync(async () =>
             {
-                var registryService = serviceProvider.GetService(typeof(IRegistryService)) as IRegistryService
-                    ?? throw new InvalidOperationException("Registry service not available");
+                var registryService = serviceProvider.GetRequiredService<IRegistryService>();
 
                 var reference = parseResult.GetValue(referenceArg)!;
                 var outputDir = parseResult.GetValue(outputOption)!;